requirements.txt
backtesting
python-binance
pandas
numpy
ta
python-dotenv
